feat(API GraphQL updateProduct): validate product category before updating a product
In alignment with the previous enhancements made to the addProduct
mutation, this commit further improves the updateProduct mutation. Now,
before updating a product, the mutation includes a validation check to
verify the existence of the provided product category in the database.

If the provided product category does not exist in the database, the
updateProduct mutation will return an appropriate error message:
"The specified category does not exist"

By adding this validation, now admins who use this mutation will get a
relevant error message instead of the vague "insert or update on table
\"products\" violates foreign key constraint
\"fk_products_product_categories\"" message.