chore(updateProduct): update GraphQL resolver error message for invalid product id
Changed the error message in the GraphQL resolver that updates a
product. The previous message was specific and included the product id
value that caused the error. To make this error easier to handle for a
client application, I changed this error to be more generic:
"A product with the specified id does not exist".