refactor(API): rename 'file-upload.ts' to 'amazon-s3.ts'
This commit renames the 'file-upload.ts' model to 'amazon-s3.ts' to
better reflect its purpose. Previously, this file was used only for
handling file uploads into an S3 bucket. However, it now includes
additional functions like "getImageUrlByObjectKey" and
"doesS3ObjectExist". Furthermore, the file may see the addition of
more functions in the future, all related to working with Amazon S3.
This is why it's appropriate to rename this file to 'amazon-s3.ts'.

All other files that depend on this module have been updated to use
appropriate import statements.