fix(API GraphQL addProduct): check magic number of S3 objects
This commit changes the way the addProduct mutation works by enhancing
its ability to recognize non-image files.

When a product is added via the addProduct mutation, the
"initialImageName" and "additionalImageName" arguments are provided.
Previously, the resolver that is associated with this mutation only
checked the availability of these 2 images in the S3 bucket. File type
checking was delegated to the way AWS handles presigned URLs uploads.
When a presigned URL was generated, it was necessary to specify which
MIME type the uploaded object has. However, when an object is uploaded
to AWS S3 via a presigned URL, S3 only checks the metadata of the
object, so technically it was possible to disguise a non-image file as
an image by simply changing the extension to .png. This way, it was
possible to upload an executable (for example) into S3.

This commit modifies the addProduct mutation by enforcing the correct
MIME type (it should be "image/png"). Whenever a user makes an API
request to the addProduct endpoint, the following will happen:
- A user will be validated. This includes verifying that the user is
authenticated, activated and has admin privileges
- The provided "initialImageName" and "additionalImageName" arguments
will be validated by checking whether they are indeed available in the
S3 bucket
- The MIME types of the provided images will be checked by verifying
their magic number (a specific sequence of bytes at the beginning of a
file that is used to identify the file's format). The MIME type must be
equal to "image/png", otherwise an error will be thrown ('The MIME type
of initialImageName is not "image/png"' or 'The MIME type of
additionalImageName is not "image/png"')
- The product category is then validated by checking whether this
category exists in the database.

If any of these checks fail (except the user validation), the provided
image names will be deleted from S3 because the product creation process
will be aborted, so there's no point in keeping the images in the S3
bucket.

Note, however, that theoretically a magic number can be forged, so this
method of validation is not foolproof, but it does provide an
additional layer of security.