feat(AWS): add a Lambda function that deletes unused S3 product images
This commit implements a lambda function that will be executed every
day at 12 AM. This function deletes all the S3 product images (from the
'onlinestore-product-images' bucket) that are not associated with any
product in the DB. These images are not needed and they only take up
space.

In theory, these unused images should not be in the S3 bucket in the
first place, but if for some reason they are there, they should be
deleted automatically now with this lambda function.

Note: if you want to change the code of this lambda function, you need
to make changes to the function codebase and then run
"npm run lambda:deploy" while being in the main lambda function folder.
For example:
user@PC:~/onlineStore$ cd aws/lambda-functions/delete-unused-s3-objects/
user@PC:~/onlineStore/aws/lambda-functions/delete-unused-s3-objects$ npm run lambda:deploy