fix(API updateProduct mutation): check MIME types of images
This commit enhances the updateProduct GraphQL mutation by checking the
MIME types of "initialImageName" and "additionalImageName" in the S3
bucket. The check is performed by checking the magic numbers of the
objects (magic numbers are the first bits of a file which uniquely
identify the type of file). If the MIME type is not 'image/png', one of
these two messages will be returned to the user depending on which
image is invalid:
1. "The MIME type of initialImageName is not \"image/png\""
2. "The MIME type of additionalImageName is not \"image/png\""

Although magic number check is not foolproof (theoretically it can be
forged), but it does provide an additional layer of security.

Note: if one or both images have invalid MIME types, they won't be
deleted from the S3 bucket ('onlinestore-product-images'). The reason
for this is that it would be impractical to take each edge case into
account. Deleting the images would generate a lot of bugs in the
future, because it's really tricky to verify if the image should be
deleted or should be left alone. This is why it's better for the images
to accumulate, and then the existing 'deleteUnusedS3Objects' Lambda
function will run every day at 12 AM to delete such unused S3 objects.

Documentation regarding the 'updateProduct' mutation was also updated
to reflect this change.