docs: update 'addProduct' mutation documentation for MIME Type Check
This commit updates the documentation around the addProduct GraphQL
mutation to add different error scenarios where the MIME type of the
uploaded images is not 'image/png'.

When the MIME type check was implemented in the addProduct mutation
(check the commit d2ae1ba), the documentation was not updated to
reflect the changes. This is why these error cases were added with this
commit. Now the documentation for this mutation is more detailed and
accurate.