fix(API addProduct mutation): delete S3 images if they are not tied to any other product
Previously, the resolver that is tied to the addProduct mutation
deleted the specified "initialImageName" and "additionalImageName" from
the S3 bucket if at least one of these conditions is met:
- If either of the provided images does not exist in the S3 bucket.
- If either of the provided images has an incorrect MIME type (anything
other than 'image/png').
- If the provided product category does not exist in the database.

For example, if you've just uploaded two images to S3 and then try to
execute the addProduct mutation with these images and you specify an
incorrect product category, these 2 images will be deleted from the S3
bucket. However, the previous implementation didn't take into account
the possibility that these images may be tied to another product.
Because of this oversight, some products may start to point to images
that don't exist because these images were deleted by the addProduct
resolver.

This commit fixes this bug by checking if the images point to another
product. If this is true, even if the addProduct request is invalid (for
example if the specified product category is incorrect), these images
will not be deleted from the S3 bucket.

Additional Note:
This fix doesn't really affect the resolver performance to the end user.
This check and the possible S3 object removal is done in the background.
So if you specified an invalid product category, the error will be
returned and the user will see it immediately. But the server will
still check in the background if the S3 objects are tied to any other
product, and if it's not the case, these S3 objects will be removed.