fix(API getMagicNumberMimeType): increase the byte range for magic numbers check
This commit changes the number of bytes that the "getMagicNumberMimeType"
function retrieves from an S3 object for checking the magic numbers to
determine the MIME type of the object.

Previously, the "getMagicNumberMimeType" function retrieved 1024 bytes,
which sometimes could lead to the EndOfStreamError, which was presumably
triggered because the number of bytes it retrieved wasn't necessary to
determine the file type. When the number of bytes was changed to 4096,
the bug was resolved and everything works as expected.