feat(API updateProduct mutation): delete S3 images if they are not tied to any other product
Previously, the updateProduct mutation didn't delete the provided S3
objects if anything went wrong in the resolver. This was done
intentionally to prevent potential bugs. The deletion of unused S3
objects was delegated to the "deleteUnusedS3Objects" Lambda function
that acts as a cron job that's executed every day at 12 AM.

However, after some consideration I decided to make the updateProduct
resolver delete the S3 objects if they are not tied to any product
(including the one we are trying to edit) AND if one of these scenarios
come into play:
- If either of the provided images does not exist in the S3 bucket.
- If either of the provided images has an incorrect MIME type (anything
other than 'image/png').
- If the provided product category does not exist in the database.

This way, if you've just uploaded two images to S3 and then try to
execute the updateProduct mutation with these images and you specify an
incorrect category, these 2 images will be deleted from the S3 bucket.
However, if these 2 images already point to any product (including the
product that you are trying to change), they will not be deleted.

The documentation was also updated to include this change.

Note, however, that the "deleteUnusedS3Objects" Lambda function will
still continue working, because it can act as a safeguard to delete any
unused S3 objects that take up unnecessary space.