docs(GraphQL): add more information about the deleteProduct mutation
This commit updates the API documentation by adding more information
about the deleteProduct GraphQL mutation. When the previous version of
the documentation was written, it didn't include all the necessary
information about the mutation. This commit aims to fix it. It fixes
typos and adds possible error responses that can occur when using this
mutation.