feat(API deleteProduct mutation): delete associated S3 images when the product is deleted
Following the addProduct and updateProduct mutations, the deleteProduct
mutation will now delete associated S3 objects as well. Now when the
product is deleted, the deleteProduct resolver will check if any other
product uses these images. If this is not the case, the images will be
deleted from the S3 bucket. Otherwise these images will stay in place.

If any error occurs and unused S3 objects remain in the S3 bucket, the
deleteUnusedS3Objects Lambda function will scan the whole bucket and
delete objects that are not associated with any product. This function
acts as a cron job and it runs every day at 12 AM.

The documentation about the deleteProduct mutation was also updated to
reflect this change.