refactor(API): move getImageName into the S3 model
Previously, the "getImageName" function was in this folder:
/server/src/graphql/helpers
However, since this function is responsible for extracting the S3 image
name from the full URL, this function would be more appropriate for the
Amazon S3 model. This model has different functions that work with
Amazon S3. This "getImageName" function was moved into the model.

This function was also renamed from "getImageName" to
"getObjectKeyByImageUrl". This new name is much more descriptive.

Different GraphQL resolvers that use this function were refactored to
use this new utility function from the Amazon S3 model.
The functionality of the resolvers wasn't changed.