refactor(React): separate button component for adding products to cart
This commit refactors the <ProductItem /> component by creating a
separate component for a button that allows you add a product to the
cart. The ProductItem component is a block that displays brief
information about a product.

The <AddToCartButton /> component is now used to separate the logic
associated with adding an item to the cart.

This promotes code modularity, readability, and improves code
maintainability.