fix(API): prevent users from adding more products to cart than available
This commit fixes the PUT /api/user/cart endpoint by checking whether
there are enough products in stock. For example, if we have 50 products
of productId = 1 and the user is trying to add 70, they won't be able
to do this, because now they will receive this message:

{
    "errors": [
        {
            "message": "insufficient stock available for this product",
            "field": "quantity"
        }
    ]
}

The previous implementation of this endpoint did not incorporate this
check, potentially allowing users to add more products to their cart
than what is currently available.

Important Note:
This fix introduces the possibility for competitors to deduce our
inventory levels by means of trial and error with API requests.
One potential mitigation for this is to impose limitations on the
quantity of products users can add to their carts. Presently, such
restrictions do not exist, but they may be enforced in the future.