fix(PUT /api/user/cart): separate productId type check and DB check
Previously, when the user hit the PUT /api/user/cart endpoint, the
'express-validator' NPM package checked the type of "productId", and
then checked the existence of such product in the database regardless
of the result of the first operation. However, if the type of the
product id is not a number, there is no point in making a request to
the database.

This commit fixes this by separating the type check and the database
check. If the productId is not a number, this message will be returned:

{
    "errors": [
        {
            "message": "productId must be a number",
            "field": "productId"
        }
    ]
}

However, if the provided product id is indeed a valid number but does
not correspond to an existing product in the database, this error
message will be returned:

{
    "errors": [
        {
            "message": "productId must point to a valid product",
            "field": "productId"
        }
    ]
}

This separation will help users pinpoint the exact error they may have
made while making an API request to add a product to the cart.