refactor(React): move <EditProduct /> to the 'pages' folder
The 'pages' folder is meant to store components that represent distinct
pages. The <EditProduct /> component fits into this category, but is
located in the 'components' folder, which is meant for small components.

This commit improves the folder structure consistency by moving the
EditProduct component to the 'pages' folder.

It's important to note that this doesn't affect the behavior of the
EditProduct component.