fix(React): extend editProductSchema from addProductSchema
This commit fixes a bug where editProductSchema and addProductSchema
weren't synchronized. The problem is that editProductSchema was almost
copy-pasted from addProductSchema, and when addProductSchema was
modified, editProductSchema remained the same.

This is why this commit changes the editProductSchema by making it
inherit properties from addProductSchema.

Now when one change is made to addProductSchema, it will propagate to
editProductSchema (except 'initialImageInfo' and 'additionalImageInfo'
because they are implemented separately in editProductSchema).