refactor(API): move validator functions into a separate folder
This commit refactors the GraphQL folder structure of the API server
code. Different resolvers use validator functions to validate the
provided arguments. Previously, these validator functions were in the
'helpers' folder. However, to add more clarity it was decided to create
a separate folder for them. This folder is called 'validators'.

Now it's easier to differentiate between different functions that are
used for handling GraphQL requests.