fix(API GraphQL adminProduct): throw an error if a product wasn't found
Previously, if the user tried to execute the adminProduct mutation with
an id that doesn't point to any product, the mutation's resolver would
return this ambiguous error: "Cannot read property 'id' of undefined".
This error simply meant that the product with the specified id wasn't
found in the database. This commit fixes this by providing a more
meaningful error message: "A product with the specified id does not
exist".