feat!(API GraphQL resolvers): add support for 'max_order_quantity'
Recently, a new field was added to the 'products' table. This field is
called 'max_order_quantity' and it specifies how many products a user
can order at a time. For example, if a product with id 50 has
max_order_quantity set to 2, a single user is allowed to order no more
than 2 instances of this product.

After the addition of this field it became necessary to enhance the API
code around GraphQL queries and mutations to support this new field.
This commit aims to do just that.

New additions:
- The addProduct mutation has a new optional (!) argument called
'max_order_quantity'. If it's specified, it must be a number between 1
and 32767 (incl. 1 and 32767). If it's not specified, the default value
of 32767 is assumed.
- The updateProduct mutation now accepts a required argument called
'max_order_quantity'. Since it's required, it must be a number between
1 and 32767.
- The adminProduct query can now return 'max_order_quantity' if it's
specified.
- Other GraphQL queries were not changed in terms of behaviour, but
they were modified in terms of TypeScript (new interfaces were added
etc).
- The documentation was modified to reflect these changes.

Note:
This commit only modifies the API codebase. This addition of the
'max_order_quantity' field can potentially break the React.js
application, so this problem will have to be addressed in the future.