fix(React.js <AddProduct />): add support for 'max_order_quantity'
With the recent addition of support for the 'max_order_quantity' field
in the API server, it has become essential to make updates to the
React.js application. Both the 'add product' and 'update product'
functionalities are broken at this point.

This commit fixes the 'Add Product' form by introducing a new input
field with the name 'maxOrderQuantity'. This input allows users to
specify the maximum number of instances of a single product that can be
ordered in a single order. Once the form is submitted, this information
is sent to the server, adding the product to the database.

Note:
Technically, the API server allows to omit 'maxOrderQuantity' when
adding a product. In this case, the default value of 32767 is assumed.
However, it was decided to make the 'maxOrderQuantity' input required
on the frontend. The reason is that it would be much easier to
implement and maintain. The default value of this input is 32767, so
the user doesn't have to specify this value manually. The API
functionality wasn't affected by this commit.