fix(React.js <EditProduct />): add support for 'max_order_quantity'
With the recent addition of support for the 'max_order_quantity' field
in the API server, it has become essential to make updates to the
React.js application. The 'add product' functionality is already fixed,
but the 'edit product' frontend functionality is still broken.

This commit fixes the 'update product' form by introducing a new input
field with the name 'maxOrderQuantity'. This input allows users to
specify the maximum number of instances of a single product that can be
ordered in a single order. This new field is required. Once the form is
submitted, this information is sent to the server, updating the product
in the database.