perf(API GraphQL queries): dynamically build SQL queries
This commit optimizes available GraphQL query resolvers to make them
dynamically build SQL queries instead of fetching everything. This is
done by getting fields that the user requested using the
'graphql-fields' library. After this the SQL query is built. Instead of
building them manually, it was decided to use a dynamic query builder
like 'knex'. This library was used only to build queries, not to
execute them. The execution is still done via the 'pg' library to
reduce the number of connections to the DB. After this the requested
fields are returned to the user and nothing more.

Benefits of dynamically building the SQL query:
- Reduced load on the database. Instead of fetching everything, now we
can be more granular.
- Reduced load on the network between the API server and the database
server. Since we are only fetching what is needed, this will reduce the
amount of unnecessary data flowing through the network.

Additional Notes:
- Query resolvers that were affected by this change:
    - getAdminProduct
    - getFeaturedProducts
    - getProduct
    - getProductsByPage
- Even though the author of the 'graphql-fields' library recommends
using 'graphql-parse-resolve-info' instead, I used the former, because
it's simple and suitable for this use case.