fix(API GET /api/user/cart): add 'initial_image_url' support
Previously, every product in the database had the 'preview_url' field.
However, now every product has the 'initial_image_url' and the
'additional_image_url' fields. The problem is that when this change
was introduced to the database schema, the /api/user/cart endpoint
wasn't modified to reflect this change. This is why the controller
that's behind this endpoint always returns an error because this
controller tries to command a model to retrieve the non-existent
'preview_url' field for each product in the cart.

This commit fixes the /api/user/cart endpoint by adding support to the
'initial_image_url' field ('additional_image_url' is not returned
because it's not needed for our use case).

Now if you want to get the contents of your cart, you have to provide
your access token in the 'Authorization' header, and the endpoint will
return you the output that adheres to this schema:

{
    "products": {
        "productId": number;
        "title": string;
        "price": number;
        "initialImageUrl": string;
        "quantity": number;
    }[]
}