fix(API order & reset-password & signup controllers): release unused DB connections
This commit fixes these controller functions:
- resetPasswordController.changePassword
- orderController.createOrder
- signupController.signup

All these functions acquired a database connection from a pool to
create a transaction. But when the transaction is no longer needed
(because the transaction is committed or a rollback is performed), the
connections weren't released back in the pool, consuming both the API
and the DB server resources. This commit fixes this by releasing the
acquired DB connections.