feat(API): add the GET /api/user/cart/count endpoint
GET /api/user/cart/count

This endpoint allows to retrieve the number of items that you have in
your cart. This endpoint doesn't require any parameters except your
access token. It's provided in the 'Authorization' header:

Authorization: Bearer your_access_token

Response:
{
    "cartItemCount": number
}

Context:
The React.js application needs a way to display the number of
cart items on the main page. It can be done by requesting a full list
of cart items (GET /api/user/cart) and then counting the items on the
client side. However, it has some clear downsides:
- More unnecessary data is flowing through the network, decreasing the
performance. We only need the quantity of items in the cart, not the
title/price of the items.
- As the application scales, server-side counting will be able to
handle the counting more efficiently than the client.
This is why implementing a separate endpoint for cart items counting is
necessary.