feat(React.js <MyCartButton />): show the number of items in the cart
When an authenticated user uses the React.js application, they will
see the 'My Cart' button at the top of the page. This button shows how
many products this user has. Previously, there was a hardcoded value of
0. However, this commit enhances this <MyCartButton /> component by
creating a request to the API server (GET /api/user/cart/count). This
endpoint returns the total number of items in the cart. This value is
then embedded in the component.

In the future, when the user is able to add new items in the cart,
some refetching strategy will need to be implemented to reflect the
current number of cart items.