fix(DELETE /api/user/cart/:productId endpoint): validate productId type
When using the '/api/user/cart/:productId' endpoint, it was necessary
to provide a number as an ID. However, technically you could provide
anything there, including a string. For example, you could make a DELETE
request to this URL: "/api/user/cart/random_string_instead_of_id"
This would return an unexpected error from the server:
{
    "errors": [
        {
            "message": "Something went wrong"
        }
    ]
}
This error was caused because the deleteFromCart controller expected
the provided product id to be a number instead of a string.

This commit fixes this by adding a new validation step to check that
the provided product id is a number. If it's not, this meaningful
message will be returned:
{
    "errors": [
        {
            "message": "productId must be a number",
            "field": "productId"
        }
    ]
}