feat(API): create a separate function to fetch product details
Added a new function to the product model to get details about an
individual product. This new function builds a SQL query dynamically.
If you specify that you want to get 'title', 'price' and nothing else,
the SQL query will only include these fields. This way, we don't have
the overfetching problem. This function will be used in the future for
REST endpoints.

There was another function called "getProductQuantity", which was also
in the product model. This function was used in the "PUT /api/user/cart"
endpoint to validate that the user isn't trying to add more products
than are available in stock. Since we now have a separate function to
fetch product details, the "getProductQuantity" function is no longer
needed, we can use the new function instead.

It was decided not to modify GraphQL queries that fetched product
details, because the modified code wouldn't be much shorter, if at all.
In addition to this, modifying the queries involves a risk of
introducing unexpected bugs, so it was decided to leave the GraphQL
queries as is.