feat(API Redis): add command timeout
Added a command timeout to the Redis client inside the API server. If a
Redis command, which was initiated from the API server, does not return
a reply within a set number of milliseconds, a "Command timed out"
error will be thrown.

This timeout is controlled with process.env.REDIS_COMMAND_TIMEOUT in
milliseconds.

This change is necessary to prevent Redis commands from hanging. A new
feature will soon be implemented within the API server that allows to
use Redis to cache certain responses from the PostgreSQL server. If the
Redis server is down, this change will prevent the API requests from
hanging.