feat(API): cache carts in Redis
This commit adds support for cart caching. The cart is cached only for
a specified amount of time. The exact time in seconds is specified in
process.env.REDIS_CART_EXPIRATION (usually it's for 1 day only).

These API endpoints were affected:

1. GET /api/user/cart
When you try to fetch the cart, it will first try to hit Redis. If the
cart isn't there, it will make a request to PostgreSQL and then cache
it in Redis. The caching process will happen asynchronously, the user
won't have to wait for the API server to cache the cart. The user will
get the response as soon as the cart is fetched from PostgreSQL.

Before implementing caching, fetching the cart would take from 40 ms to
200 ms depending on whether there are available Postgres connections in
the pool. When caching the cart with Redis, the cart loads in ~4 ms (!)
Note, however, that these values were obtained when testing the dev
environment: the Redis server is local, and the PostgreSQL server is
remote and hosted on AWS RDS which is located in Stockholm.

2. GET /api/user/cart/count
When you try to count how many products you have in your cart, the
controller behind this endpoint will try to fetch the cart from Redis
and then calculate the number of items manually using the Array.reduce
function. If the cart isn't cached, a SQL query will be sent that will
count the number of items. The response from PostgreSQL will not be
cached, however.

The performance boost is similar to the previous endpoint.

3. PUT /api/user/cart
When you try to add an item to the cart or update the quantity of an
existing cart item, the amount of time it takes to send the response
to the client would actually be higher than before, because the
controller will try to add an item to PostgreSQL and Redis, and the
client will have to wait for all these operations to finish. This is
done to avoid consistency issues.

4. DELETE /api/user/cart/:productId
When you try to delete an item from the cart, it would take longer to
send the response to the client, because the controller will try to
remove the item from both PostgreSQL and Redis. The server will send
the response only once both operations are done. Again, this is done to
avoid consistency issues.

5. POST /api/order
When you try to create an order, it would take a bit longer to finish
this because the controller behind this endpoint uses the "cleanCart"
function which empties the cart for PostgreSQL and Redis.

Note:
- It's absolutely crucial to ensure high availability of the Redis
server/cluster. Otherwise endpoints that mutate the cart may take
MUCH longer than before implementing the caching strategy. The current
Redis command timeout* is 500 ms, so if the Redis server is down,
the cart requests will take 500 ms + request to the PostgreSQL server.

*Command timeout - specifies the maximum amount of time, in
milliseconds, that the Redis client will wait for a response from the
Redis server after sending a command.