feat(API PostgreSQL): enable TLS encryption for DB communications
Previously, the communication between the API server and the PostgreSQL
RDS instance was done without the use of TLS encryption. This commit
fixes this.

To address this issue, a new parameter group was created in the AWS
account with the 'rds.force_ssl' option set to '1' (true). This option
enforces the TLS encryption on the RDS side and rejects any attempts to
initiate a non-encrypted communication.

Then the API server application code was modified to add support for
encrypted DB communications.

It's important to note that the performance of the DB queries may be
slowed down because of this change, because a TLS handshake is needed
before SQL queries are sent and executed, but this change is absolutely
necessary to ensure that the DB traffic isn't intercepted and read by
malicious actors.